+++ includes/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pda-silv <pda-silv@student.42porto.com>    +#+  +:+       +#+        */
+/*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/16 11:39:39 by pda-silv          #+#    #+#             */
-/*   Updated: 2025/04/18 18:13:35 by pda-silv         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:49:00 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -111,8 +111,12 @@ void	ft_shutdown(t_data **data, int retval);
 //	Execution
 void	ft_execute(t_data *data);
 char	*ft_getenv(const char *name, char **env);
+void	ft_execute_command(t_data *data, char **cmd_args, t_token_type type);
 
 //	Builtin
+void	ft_execute_lone_builtin(t_data *data, int cmd_index, char **cmd_args);
+void	ft_handle_command(t_data *data, int *pipefd, int cmd_index,
+			char **cmd_args);
 void	ft_execute_builtin(t_data *data, char **cmd_args);
 void	ft_exit(t_data *data, char **cmd_args);
 void	ft_print_sorted_env(char **env);
@@ -128,7 +132,8 @@ void	ft_getpath(t_data *data, char *arg);
 char	*ft_expand_variables(char *word, t_data *data);
 
 //	Pipe
-void	ft_handle_pipes(t_data *data, int pipefd[2], t_command command, int cmd_index);
+void	ft_handle_pipes(t_data *data, int pipefd[2],
+			t_command command, int cmd_index);
 void	ft_wait_children(t_data *data, pid_t *pids);
 void	ft_pipe_error(t_data *data, char	**cmd_args);
 
@@ -142,4 +147,20 @@ void	ft_open_redirect_fds(t_redir *redir, const char *in, const char *out);
 void	ft_close_redirect_fds(t_redir *redir);
 void	ft_get_delim_buf(t_command *command, char *delim);
 
+// Sorting
+
+void	ft_handle_pipes(t_data *data, int pipefd[2],
+			t_command command, int cmd_index);
+void	ft_handle_heredoc(t_data *data, t_command cmd, int cmd_index);
+char	*ft_remove_quotes(char *str);
+int		handle_quotes(char **ptr, int *in_quotes, char *quote_type);
+char	*handle_quote(char *word, int *i, char *result, int *in_quotes);
+char	*append_char(char *str, char c);
+void	ft_tokenize_input(t_data *data);
+void	ft_free_cmd(t_data *data, char **cmd_args);
+char	*ft_skip_whitespace(char *ptr);
+t_token	ft_parse_token(char **ptr, t_data *data);
+int		ft_handle_quotes_in_input(t_data *data);
+void	swap_strings(char **a, char **b);
+int		var_exists(char **env, char *var);
 #endif
