+++ src/builtins/echo.c
@@ -6,7 +6,7 @@
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 14:27:48 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/05 10:33:31 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:38:24 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,14 +31,15 @@ int	is_valid_echo_flag(const char *flag)
 /**********************/
 /*Echo Builtin Command*/
 /**********************/
-void ft_echo(char **cmd_args)
+void	ft_echo(char **cmd_args)
 {
-	int newline;
-	int i;
+	int	newline;
+	int	i;
 
 	newline = 1;
 	i = 1;
-	while (cmd_args[i] && cmd_args[i][0] == '-' && is_valid_echo_flag(cmd_args[i]))
+	while (cmd_args[i] && cmd_args[i][0] == '-'
+		&& is_valid_echo_flag(cmd_args[i]))
 	{
 		newline = 0;
 		i++;
