+++ src/builtins/exit.c
@@ -6,7 +6,7 @@
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 14:27:48 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/05 10:33:29 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:36:58 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
