+++ src/builtins/export.c
@@ -6,7 +6,7 @@
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 14:27:48 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/05 11:06:28 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:44:36 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,23 +73,6 @@ static void	process_var_with_equal(t_data *data, char *var)
 			j++;
 		add_env_variable(data, var, j);
 	}
-}
-
-static int	var_exists(char **env, char *var)
-{
-	int	i;
-	int	var_len;
-
-	i = 0;
-	var_len = ft_strlen(var);
-	while (env[i])
-	{
-		if (ft_strncmp(env[i], var, var_len) == 0 &&
-			(env[i][var_len] == '=' || env[i][var_len] == '\0'))
-			return (1);
-		i++;
-	}
-	return (0);
 }
 
 static void	process_var_no_equal(t_data *data, char *var)
