+++ src/builtins/export_supp.c
@@ -1,21 +1,39 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pwd.c                                              :+:      :+:    :+:   */
+/*   export_supp.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 14:27:48 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/22 21:36:22 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:44:43 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-/*********************/
-/*PWD Builtin Command*/
-/*********************/
-void	ft_pwd(t_data *data)
+void	swap_strings(char **a, char **b)
 {
-	ft_printf("%s\n", data->cwd);
+	char	*temp;
+
+	temp = *a;
+	*a = *b;
+	*b = temp;
+}
+
+int	var_exists(char **env, char *var)
+{
+	int	i;
+	int	var_len;
+
+	i = 0;
+	var_len = ft_strlen(var);
+	while (env[i])
+	{
+		if (ft_strncmp(env[i], var, var_len) == 0
+			&& (env[i][var_len] == '=' || env[i][var_len] == '\0'))
+			return (1);
+		i++;
+	}
+	return (0);
 }
