+++ src/builtins/unset.c
@@ -6,7 +6,7 @@
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 14:27:48 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/05 10:33:23 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:38:00 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,11 +47,12 @@ static int	is_matching_var(char *env_var, char *arg)
 	var_len = 0;
 	while (arg[var_len] && arg[var_len] != '=')
 		var_len++;
-	if (ft_strncmp(env_var, arg, var_len) == 0 &&
-		(env_var[var_len] == '=' || env_var[var_len] == '\0'))
+	if (ft_strncmp(env_var, arg, var_len) == 0
+		&& (env_var[var_len] == '=' || env_var[var_len] == '\0'))
 		return (1);
 	return (0);
 }
+
 /*********************/
 /*Seach Env to Update*/
 /*********************/
