+++ src/mario.c
@@ -6,13 +6,13 @@
 /*   By: joseferr <joseferr@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/04 13:30:00 by joseferr          #+#    #+#             */
-/*   Updated: 2025/04/22 20:39:39 by joseferr         ###   ########.fr       */
+/*   Updated: 2025/04/22 21:31:51 by joseferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void ft_wait_children(t_data *data, pid_t *pids)
+void	ft_wait_children(t_data *data, pid_t *pids)
 {
 	int	i;
 	int	status;
@@ -39,7 +39,7 @@ void ft_wait_children(t_data *data, pid_t *pids)
 /*   Sets up output redirections to files or pipes                           */
 /*   Ensures proper synchronization of heredoc processing                    */
 /* ************************************************************************** */
-static void	ft_handle_heredoc(t_data *data, t_command cmd, int cmd_index)
+void	ft_handle_heredoc(t_data *data, t_command cmd, int cmd_index)
 {
 	int	heredoc_pipe[2];
 
@@ -62,73 +62,6 @@ static void	ft_handle_heredoc(t_data *data, t_command cmd, int cmd_index)
 		dup2(heredoc_pipe[0], STDIN_FILENO);
 		close(heredoc_pipe[0]);
 	}
-}
-
-/* ************************************************************************** */
-/*                                                                            */
-/*   Handles input redirection setup for commands                            */
-/*   Processes input from files or previous pipes                            */
-/*   Ensures correct file descriptor duplication                             */
-/*   Properly closes original file descriptors after duplication             */
-/* ************************************************************************** */
-static void	ft_handle_input(t_data *data, t_command cmd, int cmd_index)
-{
-	if (cmd.redir.in_fd != STDIN_FILENO)
-	{
-		dup2(cmd.redir.in_fd, STDIN_FILENO);
-		close(cmd.redir.in_fd);
-	}
-	else if (cmd_index > 0)
-	{
-		dup2(data->prev_pipe, STDIN_FILENO);
-		close(data->prev_pipe);
-	}
-}
-
-/* ************************************************************************** */
-/*                                                                            */
-/*   Handles output redirection setup for commands                           */
-/*   Manages output to files or next pipe in the pipeline                    */
-/*   Ensures correct file descriptor duplication                             */
-/*   Properly closes original file descriptors after duplication             */
-/* ************************************************************************** */
-static void	ft_handle_output(t_data *data, t_command cmd,
-							int pipefd[2], int cmd_index)
-{
-	if (cmd.redir.out_fd != STDOUT_FILENO)
-	{
-		dup2(cmd.redir.out_fd, STDOUT_FILENO);
-		close(cmd.redir.out_fd);
-	}
-	else if (cmd_index < data->cmd_count)
-	{
-		dup2(pipefd[1], STDOUT_FILENO);
-		close(pipefd[1]);
-	}
-}
-
-/* ************************************************************************** */
-/*                                                                            */
-/*   Coordinates pipe and redirection handling                               */
-/*   Sets up inheritance of file descriptors between commands                */
-/*   Processes heredoc input when needed                                     */
-/*   Ensures proper synchronization in the pipeline                          */
-/* ************************************************************************** */
-void	ft_handle_pipes(t_data *data, int pipefd[2],
-						t_command command, int cmd_index)
-{
-	if (cmd_index > 0)
-	{
-		char	dummy;
-
-		read(data->heredoc_sync[cmd_index - 1][0], &dummy, 1);
-		close(data->heredoc_sync[cmd_index - 1][0]);
-	}
-	if (command.redir.delim)
-		ft_handle_heredoc(data, command, cmd_index);
-	else
-		ft_handle_input(data, command, cmd_index);
-	ft_handle_output(data, command, pipefd, cmd_index);
 }
 
 void	ft_setup_pipes(int pipefd[2])
